fix: resolve all remaining compilation errors
- Disambiguate ContentMap import (hide from analysis_service.dart)
- Fix Appwrite listDocuments queries parameter type
- Fix ProgressCallback signature (add String stage param)
- Remove non-existent CaptionPosition.left/.right from switches
- Fix editor_providers split clip (use TimelineClip constructor)
- Add toJson() to EditDecision class
- Rename LongFormListNotifier/ChapterListNotifier.update() to avoid Riverpod conflict
- Make ProjectRepository cacheTtl nullable
- Add dio import for FormData/MultipartFile in upload_service
- Fix DateTime reference-before-declaration in date_time_extensions
- Fix clamp() return type (add .toDouble())
- Add ProGuard rule for Google Play Core missing class
- Fix const DateTime issues in test file

+++ lib/shared/extensions/double_extensions.dart
@@ -80,7 +80,7 @@ extension DoubleExtensions on double {
   /// 150.0.clampValue(0, 100); // 100.0
   /// 50.0.clampValue(0, 100); // 50.0
   /// ```
-  double clampValue(double min, double max) => clamp(min, max);
+  double clampValue(double min, double max) => clamp(min, max).toDouble();
 
   /// Returns true if the value is approximately equal to another value.
   ///
